% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisquare_extractor.R
\name{extractChiSquare}
\alias{extractChiSquare}
\title{Extract  chi square tests from text
extractTestStats <- function(inputText, context = FALSE, contextSize = 100, sectionName = NA)}
\usage{
extractChiSquare(input)
}
\arguments{
\item{input}{input text}
}
\value{
A tibble contaitning all detected  chi square tests, 
associated degrees of freedom and p values
}
\description{
Extract  chi square tests from text
extractTestStats <- function(inputText, context = FALSE, contextSize = 100, sectionName = NA)
}
\examples{
extractChiSquare("chi2(123, n = 12) = .01, p = > .05")


}
