% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_extractor.R
\name{extractCorrelations}
\alias{extractCorrelations}
\title{Extract correlation coefficents from text
extractTestStats <- function(inputText, context = FALSE, contextSize = 100, sectionName = NA)}
\usage{
extractCorrelations(input)
}
\arguments{
\item{inputText}{input text}

\item{contextSize}{= size of context around extracted test statistics}
}
\value{
A tibble contaitning all detected correlation coefficents, 
associated degrees of freedom and p values, with context of size contextSize.
}
\description{
Extract correlation coefficents from text
extractTestStats <- function(inputText, context = FALSE, contextSize = 100, sectionName = NA)
}
\examples{
extractCorrelations("r(123) = .01, p = < .001")


}
