% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftest_extractor.R
\name{extractFTests}
\alias{extractFTests}
\title{Extract F tests from text
extractTestStats <- function(inputText, context = FALSE, contextSize = 100, sectionName = NA)}
\usage{
extractFTests(input)
}
\arguments{
\item{input}{input text}
}
\value{
A tibble contaitning all detected F tests, 
associated degrees of freedom and p values
}
\description{
Extract F tests from text
extractTestStats <- function(inputText, context = FALSE, contextSize = 100, sectionName = NA)
}
\examples{
extractFTests("F(1, 23) = .01, p = < .001")


}
