% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_extractor.R
\name{extractTTests}
\alias{extractTTests}
\title{Extract t tests from text}
\usage{
extractTTests(input)
}
\arguments{
\item{inputText}{input text}

\item{contextSize}{= size of context around extracted test statistics}
}
\value{
A tibble contaitning all detected ttest coefficents, 
associated degrees of freedom and p values, with context of size contextSize.
}
\description{
Extract t tests from text
}
\examples{
extractTTests("t(123) = .01, p = .001")


}
