% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisquare_extractor.R
\name{extractChiSquare}
\alias{extractChiSquare}
\title{Extract  chi square tests from text}
\usage{
extractChiSquare(input)
}
\arguments{
\item{inputText}{input text}
}
\value{
A tibble containing all extracted values, with columns identifying the type of statistic extracted, the reported test as reported in the text, the degrees of freedom, the p value and the value reported.
}
\description{
Extract  chi square tests from text
}
\examples{
extractChiSquare("chi2(123, n = 12) = .01, p = > .05")


}
